conversion fix

+++ Project1/Project1/SoundCast.h
@@ -33,11 +33,11 @@ typedef struct _Ray
 	int fadeStrength;
 	bool trackToSource;
 	int bounces;
-	int velMult;
+	float velMult;
 	bool isPlayer;
 } Ray;
 void InitScene(Wall* _walls, int _maxWalls);
 
 void RayUpdate(float wx, float wy);
-void CreateRay(float x, float y, int length, float velx, float vely,int fade, CP_Color color, bool trackToSource, int velMult, bool isPlayer);
+void CreateRay(float x, float y, int length, float velx, float vely,int fade, CP_Color color, bool trackToSource, float velMult, bool isPlayer);
